見た目をマシにしました(まだ気持ち悪い)

+++ api/schemas/sleepdata.py
@@ -7,7 +7,7 @@ from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
 from jwt.exceptions import InvalidTokenError
 from passlib.context import CryptContext
 from pydantic import BaseModel, Field, ConfigDict
-from sqlalchemy import create_engine, Column, String, Integer, Boolean, ForeignKey
+from sqlalchemy import create_engine, Column, String, Integer, Float, Boolean, ForeignKey
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker, Session, Mapped, relationship
 from api.database import SessionLocal, engine, Base
@@ -23,7 +23,7 @@ class SleepDataInDB(Base):
     date = Column(String) #2024-07-12
     bedtime = Column(String) #2024-07-12T22:03
     wakeup = Column(String) #2024-07-12T22:03
-    sleeptime = Column(Integer) #9
+    sleeptime = Column(Float) #9
     #weekday = Column(Integer) #月0~日6
 
 class SleepBase(BaseModel):
@@ -31,7 +31,7 @@ class SleepBase(BaseModel):
     date: str | None = Field(None, description="睡眠日時(就寝日)")
     bedtime: str | None = Field(None, description="就寝時間")
     wakeup: str | None = Field(None, description="起床時間")
-    sleeptime: int | None = Field(None, description="睡眠時間")
+    sleeptime: float | None = Field(None, description="睡眠時間")
 
 class SleepData(SleepBase):
     id: int | None = Field(None, description="プライマリーキー")
